Added and setup HTML, CSS, and JS files for initial structure and logic
HTML file is given a title and some divs for the basic structure.

CSS file changes the bg color and sets up the sizes, padding, and margins of the squares

JS file hard codes the squares and the sqaure that should be picked with RGB values of my choice.  It also implements logic to see if the color picked matches the color clicked.